requirements.txt
pyqlib>=0.9.0
pandas>=1.5.0
numpy>=1.23.0
scipy>=1.9.0
matplotlib>=3.6.0
tqdm>=4.64.0
pyarrow>=10.0.0